+++ how-to/workspace-platform-starter/client/src/framework/shapes/menu-shapes.ts
@@ -23,6 +23,12 @@ export interface MenusProviderOptions extends ModuleList {
 	popupHtml?: string;
 
 	/**
+	 * The font size used in the custom popup menu.
+	 * defaults to 12
+	 */
+	menuFontSize?: number;
+
+	/**
 	 * The width to display the custom popup menu.
 	 * defaults to 200
 	 */
@@ -204,31 +210,11 @@ export type MenuOptionType<T> = T extends GlobalContextMenuItemTemplate
 	: TrayMenuOptionType;
 
 /**
- * Entry for a popup menu.
+ * The styles that can be used to display the popup menus.
  */
-export interface PopupMenuEntry<T = unknown> {
-	/**
-	 * Type of the menu entry.
-	 */
-	type?: "normal" | "separator";
+export type PopupMenuStyles = "native" | "custom";
 
-	/**
-	 * Label to show in the menu.
-	 */
-	label?: string;
-
-	/**
-	 * Icon to show in the menu.
-	 */
-	icon?: string;
-
-	/**
-	 * Custom data to associate with the entry.
-	 */
-	customData?: T;
-
-	/**
-	 * Is the menu entry enabled.
-	 */
-	enabled?: boolean;
-}
+/**
+ * Specialized version of the menu item template with generic data.
+ */
+export type PopupMenuEntry<T = unknown> = Omit<OpenFin.MenuItemTemplate, "data"> & { data?: T };